Update feat_wild_talent.py (#17)
* Update feat_wild_talent.py

Made some corrections

Should Wild Talent affect the Base Max Psi too? For purposes of determining feat eligibility and such.

Couple of notes:
You do not use Constructors for PythonModifiers! The only thing a PythonModifier's constructor will do is register it in the grand list of modifiers, that's it.

You can add a hook for ET_OnConditionAdd which will go off whenever a modifier is isntantiated onto a specific object. But even there you should be very careful!

* Rename feat_wild_talent.py to scr/tpModifiers/feat_wild_talent.py

* Rename scr/tpModifiers/feat_wild_talent.py to feat_wild_talent.py

* Changed to affect Base Max Psi and renamed